backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
openai==1.30.1
sentence-transformers==2.7.0
faiss-cpu==1.8.0
numpy==1.26.4
motor==3.4.0
pymongo==4.7.2
httpx==0.27.0
